requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
asyncpg>=0.29
pydantic-settings>=2.2
redis>=5.0
alembic>=1.13
psycopg[binary]>=3.1
email-validator>=2.1
PyJWT>=2.8
